refactor(#790): promote format_local_short to src/display_time.rs
§3.10 prep: introduce shared display-time helper and FleetConfig
schema field. Behaviour preserved — tz param ignored in this commit,
delegate falls through to chrono::Local exactly like Sprint 54 P2-6
panels.rs version. Next commits add the IANA tz path (anchor RED then
GREEN) and wire ci-watch notification body sites (sweep.rs).

- New `src/display_time.rs` with `format_local_short(rfc3339, tz)` —
  delegate impl; 3 re-hosted tests from panels.rs cover Z-form shape,
  explicit-offset equivalence, parse-error fallback
- Add `FleetConfig.display_timezone: Option<String>` field
  (additive, `skip_serializing_if` keeps existing fleet.yaml clean)
- Rewire `panels.rs:64` decisions panel render to call the shared
  helper with `None` (preserves chrono::Local fallback)
- Delete duplicate `format_local_short` def + 3 dup tests from
  panels.rs after re-host
- Plumb `display_timezone: None` into fleet_normalize.rs test fixture

α scope per dispatch spec d-20260514191344776044-7: storage stays UTC
unconditionally; only human-facing surfaces convert. Machine-API
responses (inbox, list_instances, task/decision JSON) stay UTC.


Agend-Agent: fixup-dev
Agend-Task: self
Agend-Branch: fix/790-timezone-display-config
Agend-Issued-At: 2026-05-14T19:14:27.866890+00:00
